+++ src/main/java/com/jamjam/bookjeok/domains/payment/dto/response/PaymentConfirmResponse.java
@@ -1,6 +1,6 @@
 package com.jamjam.bookjeok.domains.payment.dto.response;
 
-import com.jamjam.bookjeok.domains.order.dto.orderdetail.OrderDetailDTO;
+import com.jamjam.bookjeok.domains.orderdetail.dto.OrderDetailDTO;
 import lombok.Builder;
 
 import java.util.List;
